requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.5.0
django-email-utils==1.0.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
